requirements.txt
opencv-python
mediapipe
numpy
SpeechRecognition
pyaudio
pydub
openai
python-dotenv